chore: reafctor agent sdk
