+++ packages/agent-sdk/bunup.config.ts
@@ -4,6 +4,9 @@ export default defineConfig({
   entry: ['src/index.ts'],
   outDir: 'dist',
   format: ['esm', 'cjs'],
-  dts: true,
+  dts: {
+    resolve: true,
+  },
   clean: true,
+  noExternal: ['@browseros/shared'],
 })
