+++ packages/agent-sdk/src/types.ts
@@ -1,34 +1,9 @@
+import type { LLMConfig, LLMProvider } from '@browseros/shared/schemas/llm'
+import type { UIMessageStreamEvent } from '@browseros/shared/types/ui-stream'
 import type { ZodSchema } from 'zod'
 
-/**
- * Supported LLM providers for browser automation
- */
-export type LLMProvider =
-  | 'anthropic'
-  | 'openai'
-  | 'google'
-  | 'openrouter'
-  | 'azure'
-  | 'ollama'
-  | 'lmstudio'
-  | 'bedrock'
-  | 'browseros'
-  | 'openai-compatible'
-
-/**
- * LLM configuration for agent operations
- */
-export interface LLMConfig {
-  provider: LLMProvider
-  model?: string
-  apiKey?: string
-  baseUrl?: string
-  resourceName?: string
-  region?: string
-  accessKeyId?: string
-  secretAccessKey?: string
-  sessionToken?: string
-}
+// Re-export shared types for consumers (bundled at build time)
+export type { LLMConfig, LLMProvider, UIMessageStreamEvent }
 
 /**
  * Configuration options for creating an Agent instance.
@@ -104,37 +79,6 @@ export interface ProgressEvent {
   /** Additional metadata about the operation */
   metadata?: Record<string, unknown>
 }
-
-/**
- * UI Message Stream events (Vercel AI SDK format).
- * These events are forwarded from the /chat endpoint during act() calls.
- */
-export type UIMessageStreamEvent =
-  | { type: 'start'; messageId?: string }
-  | { type: 'start-step' }
-  | { type: 'text-start'; id: string }
-  | { type: 'text-delta'; id: string; delta: string }
-  | { type: 'text-end'; id: string }
-  | { type: 'reasoning-start'; id: string }
-  | { type: 'reasoning-delta'; id: string; delta: string }
-  | { type: 'reasoning-end'; id: string }
-  | { type: 'tool-input-start'; toolCallId: string; toolName: string }
-  | { type: 'tool-input-delta'; toolCallId: string; inputTextDelta: string }
-  | {
-      type: 'tool-input-available'
-      toolCallId: string
-      toolName: string
-      input: unknown
-    }
-  | { type: 'tool-output-available'; toolCallId: string; output: unknown }
-  | { type: 'tool-input-error'; toolCallId: string; errorText: string }
-  | { type: 'tool-output-error'; toolCallId: string; errorText: string }
-  | { type: 'source-url'; sourceId: string; url: string; title?: string }
-  | { type: 'file'; url: string; mediaType: string }
-  | { type: 'error'; errorText: string }
-  | { type: 'finish-step' }
-  | { type: 'finish'; finishReason: string; messageMetadata?: unknown }
-  | { type: 'abort' }
 
 /**
  * Result returned by `nav()`.
